+++ docs/history.rst
@@ -6,7 +6,8 @@ Change Log
 * ENH: Added :meth:`pyproj.proj.Proj.get_factors` (issue #503) 
 * ENH: Added type hints (issue #369)
 * BUG: Don't use CRS classes for defaults in CRS child class init signatures (issue #554)
-* ENH: Updated :attr:`pyproj.crs.CRS.axis_info` to pull all relevant axis information from CRS.
+* ENH: Updated :attr:`pyproj.crs.CRS.axis_info` to pull all relevant axis information from CRS (issue #557)
+* ENH: Added :meth:`pyproj.transformer.Transform.__eq__` (issue #559)
 
 2.5.0
 ~~~~~